refactor: general circularityCheckers file
