+++ packages/core/src/fetchers/fetchById.ts
@@ -3,7 +3,7 @@ import { fetchExperienceEntry } from './fetchExperienceEntry';
 import { fetchReferencedEntities } from './fetchReferencedEntities';
 import { ExperienceEntry } from '@/types';
 import { ContentfulClientApi, Entry } from 'contentful';
-import { removeCircularReferences } from './shared/removeCircularReferences';
+import { removeCircularPatternReferences } from './shared/circularityCheckers';
 
 const errorMessagesWhileFetching = {
   experience: 'Failed to fetch experience',
@@ -57,7 +57,7 @@ export async function fetchById({
       throw new Error(`No experience entry with id: ${id} exists`);
     }
 
-    removeCircularReferences(experienceEntry as ExperienceEntry);
+    removeCircularPatternReferences(experienceEntry as ExperienceEntry);
 
     try {
       const { entries, assets } = await fetchReferencedEntities({
